backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
django-filter>=24.2,<25.0
celery>=5.4,<6.0
redis>=5.0,<6.0
psycopg[binary]>=3.2,<4.0
dj-database-url>=2.2,<3.0
gunicorn>=22.0,<23.0
boto3>=1.34,<2.0
django-storages>=1.14,<2.0
Pillow>=10.4,<11.0
